// src/models/IUser.ts
import IInputProp from "./IInputProp";

interface IUser {
    id: IInputProp;
    names: IInputProp;
    lastnames: IInputProp;
    email: IInputProp;
    phone: IInputProp;
    company: IInputProp;
    rol: IInputProp;
    commentaries: IInputProp;
    password: IInputProp;
}

export default IUser;